Fix: server error
Add basic login error messages
